file: separate folder for tf tests
