+++ bash_scripts/tf_tests.sh
@@ -57,10 +57,10 @@ N_CPU=$(grep -c ^processor /proc/cpuinfo)
 
 if [ "$integration" = "true" ]; then \
     # Run all tests
-    pytest -n "${N_CPU}" tests ;
+    pytest -n "${N_CPU}" tests/tf ;
 else
     # Run all unit tests (non integration tests).
-    pytest --durations=10 -n "${N_CPU}" tests --ignore-glob="*/*system_test.py" ;
+    pytest --durations=10 -n "${N_CPU}" tests/tf --ignore-glob="*/*system_test.py" ;
 fi
 
 # Clean-up.
